Move tests into core
